Change mip level padding to match draft 18 rules. (#170)
* Validate mip padding as specified in KTX2 draft 18.

* Create & read files with KTX2 draft 18 mip padding.

* Regen all the reference ktx2 files.

Also contains the following:

* Use info from dfd instead of vkGetFormatSize.

* Create output file for each input file - bug fix in ktx2ktx2.

* Improve validation. Fixes #151. Fixes #152.

+++ tests/transcodetests/transcodetests.cc
@@ -103,7 +103,7 @@ bool read_file( string path, void** data, long *fsize ) {
     fseek(f, 0, SEEK_SET);  /* same as rewind(f); */
 
     *data = malloc(*fsize);
-    long numRead = fread(*data, 1, *fsize, f);
+    size_t numRead = fread(*data, 1, *fsize, f);
     fclose(f);
     return numRead == *fsize;
 }
